[Exp PyROOT] Build ROOT with multiple Python versions (1)
The basic idea is to build and install all the libraries related to PyROOT
inside a directory /lib/python{version} instead of /lib, where they
currently are.
In this way, it is possible to build libraries for multiple Python
versions, then choose the one we want to use through an option in
thisroot.sh.
In this first commit, we set the variables necessary to build the Python
related libraries in the specific directory and we add a symlink to the
latter when ROOT is built for the first time.