[Exp PyROOT] Build ROOT with multiple Python versions (1/4)
The basic idea is to build and install all the libraries related to PyROOT
inside a directory
CMAKE_LIBRARY_OUTPUT_DIRECTORY/lib/python{version}
instead of
CMAKE_LIBRARY_OUTPUT_DIRECTORY/lib
where they currently are.
In this way, it is possible to build libraries for multiple Python
versions, then choose the one we want to use through an option in
thisroot.sh.
What is done here:
- move the research of Python installation in RootBuildOptions.cmake
- set a varible "py_localruntimedir" to build in the correct path
- add a symlink created only if ROOT is built for the first time