[Exp PyROOT] Build ROOT with multiple Python versions (2/4)
Here we make the necessary changes to allow the user to choose a specific
PyROOT build directory when source thisroot.sh.
Source a specific Python version by calling:

$ PYTHON_VERSION={version_number} source bin/thisroot.sh

If PYTHON_VERSION is not provided, the default version (the one located in
lib/python_default) is sourced.
If the PYTHON_VERSION is not valid (i.e. the corrispondent build does not
exist) a message is printed and no action is performed.