[Exp PyROOT] Automatically copy Python file if modified
At the moment, the build system does not detect if a Python file has
changed and the only way to apply the changes is repeat the
configuration step too (or manually copy the Python file in the build
directory).
With this change, by simply retyping "cmake --build .", also the changes
to Python files are detected and the file itself is properly changed in
the build directory.

N.B.: feature introduced only for the Python files in the PyROOT module,
not in Cppyy.