[CMake][Exp PyROOT] Set variables for both Python2 and Python3 if CMake version >= 3.14
A change is needed here in order to build PyROOT with two versions at
the same time. We still look for a "preferred" Python version based on
what "python" points to, in order to install the !!DEPRECATED!! CMake
variables PYTHON_EXECUTABLE, PYTHON_INCLUDE_DIRS ecc. used in other
parts of the software.
At the end of the search for Python 2 and 3, a series of lists
containing the pairs of Python CMake variables useful for MultiPython is
defined.