[CMake][Exp PyROOT] Remove QUIET from find_package Python2/3 and Numpy
Contrary to what was pointed out in 89c089, for what concerns
find_package(Python2), find_package(Python3), find_package(Python2
COMPONENTS Numpy) and find_package(Python3 COMPONENTS Numpy) the default
behavior is the one expected: if neither REQUIRED and QUIET are
specified, an output is printed and the build continues even if the
package is not found. For this reason it seems safe to remove the option
QUIET from here.
The behavior observed in 89c089 is still observed for what concerns
find_package(Python COMPONENTS Numpy), but this is erased in this PR.